Debug product controller and unit of work

+++ BulkyBookWeb/Areas/Admin/Controllers/ProductController.cs
@@ -48,31 +48,35 @@ public class ProductController : Controller
             return View(productVm);
         }
     }
-    
+
     //POST
     [HttpPost]
     [ValidateAntiForgeryToken]
-    public IActionResult Upsert(ProductVm obj, IFormFile file)
+    public IActionResult Upsert(ProductVm obj, IFormFile? file)
     {
         if (ModelState.IsValid)
         {
             string wwwRootPath = _hostEnvironment.WebRootPath;
-            if (file != null)
+            if (file != null || obj.Product.ImageUrl != null)
             {
-                string fileName = Guid.NewGuid().ToString();
-                var uploads = Path.Combine(wwwRootPath, @"images/products/");
-                var extension = Path.GetExtension(file.FileName);
-                if (obj.Product.ImageUrl != null)
+                if (file!= null)
                 {
-                    var oldImagePath = Path.Combine(wwwRootPath, obj.Product.ImageUrl.TrimStart('\\'));
-                    if (System.IO.File.Exists(oldImagePath)) {
-                        System.IO.File.Delete(oldImagePath);
+                    string fileName = Guid.NewGuid().ToString();
+                    var uploads = Path.Combine(wwwRootPath, @"images/products");
+                    var extension = Path.GetExtension(file.FileName);
+                    if (obj.Product.ImageUrl != null)
+                    {
+                        var oldImagePath = Path.Combine(wwwRootPath, obj.Product.ImageUrl.TrimStart('\\'));
+                        if (System.IO.File.Exists(oldImagePath)) {
+                            System.IO.File.Delete(oldImagePath);
+                        }
                     }
-                }
-                using (var fileStreams =
-                       new FileStream(Path.Combine(uploads, fileName + extension), FileMode.Create))
-                {
-                    file.CopyTo(fileStreams);
+                    using (var fileStreams =
+                           new FileStream(Path.Combine(uploads, fileName + extension), FileMode.Create))
+                    {
+                        file.CopyTo(fileStreams);
+                    }
+                    obj.Product.ImageUrl = @"images/products/" + fileName + extension;
                 }
                 if (obj.Product.Id == 0)
                 {
@@ -82,7 +86,6 @@ public class ProductController : Controller
                 {
                     _unitOfWork.Product.Update(obj.Product);
                 }
-                obj.Product.ImageUrl = @"images/products/" + fileName + extension;
             }
             _unitOfWork.Save();
             TempData["success"] = "Product created successfully";
@@ -90,7 +93,6 @@ public class ProductController : Controller
         }
         return View(obj);
     }
-    
     public IActionResult Index()
     {
         return View();
